requirements.txt
Appium-Python-Client==4.2.1
behave==1.2.6
selenium~=4.25.0